add location tests too
