+++ src/backend/common/models/tests/conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def auto_add_ndb_context(ndb_context) -> None:
+    pass
